[New] Added new test cases for referrals
